Code file
This file contains the code in the python language on the jupyter notebook. In this project two methods of sentimental analysis is covered i.e. by using any dataset and by just entering any input in the form of a sentence. Sentimental analysis can be done on both type of data.